mypy: fix remaining in-scope errors across 20 files (~32 -> 0)
Batch of the last errors in my scope after the earlier file-by-file passes. Mostly
mechanical: implicit-Optional defaults widened to Optional[T]; var-annotated dict/list
literals given explicit types (kernel_tuning_cache _parse_dims, models/selection.py
group_dict, inference/predict.py models, votenrank/stability_exp.py final_corrs);
no-any-return sites wrapped in float()/str() (utils/_param_oracle._rss_mb,
utils/disk_cache hash functions after retyping _hasher, metrics/_core_auc_brier
fast_numba_aucs caller, metrics/_multilabel_metrics._popcount64 annotation fixed to
match its actual `int(...)` body).

Real bugs fixed:
- utils/text.py read_glove_embeddings: fpath had a non-None string default but was
  typed Optional[str], which would let a caller explicitly pass None into open();
  narrowed to str (grepped all callers, none pass None).
- estimators/early_stopping_monotonic.py: added an assert that self.prev is set
  whenever self.best is (they're always assigned together in __init__ and in the
  first-observation branch), fixing a real Optional-comparison gap mypy caught.
- core/composite_similarity.py: classification was assigned a numpy bool from
  `uniq.shape[0] <= 10 and np.all(...)` into a bool-typed field; wrapped in bool().
- utils/_param_oracle.py: per_bucket keyed by r.get("fp_bucket_json") which can be
  None; defaulted to "" instead of letting a None key silently diverge from the
  dict[str, ...] contract. _metric_name() now asserts the __init__ invariant that
  exactly one of minimize/maximize is non-None rather than silently returning None.
- votenrank/fairness_computation.py: extended the type: ignore on the optional
  transformers/torch import fallback to cover the "Cannot assign to a type" [misc]
  error alongside [assignment] on both AutoTokenizer and torch.
- reporting/report_html.py: by_section's ChartEntry slot can genuinely be None
  (malformed entries); widened the annotation instead of the arg-type errors it was
  masking, and defaulted html.escape(err) to "unknown error" for the mypy-flagged
  Optional[str] case (matches the existing "(malformed)" fallback pattern nearby).

+++ src/mlframe/metrics/_multilabel_metrics.py
@@ -169,7 +169,7 @@ def _fast_jaccard_score_par(y_true: np.ndarray, y_pred: np.ndarray) -> float:
 
 
 @numba.njit(**NUMBA_NJIT_PARAMS)
-def _popcount64(x: np.uint64) -> np.int64:
+def _popcount64(x: np.uint64) -> int:
     """Population-count for uint64 -- Hacker's Delight bit-twiddle.
 
     Numba doesn't expose ``__builtin_popcountll`` directly; this 5-instruction
